ai_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
easyocr==1.7.0
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
websockets==11.0.3
asyncio==3.4.3
nfl-data-py==0.3.0
pandas==2.0.3
